Pass savedBooks props to Main and Bookshelf rendered by React Router
