+++ src/components/Bookshelf.js
@@ -2,16 +2,10 @@ import React, { Component } from 'react';
 import Cards from './Cards';
 import { Route } from 'react-router-dom';
 
-// Clear localStorage
-// localStorage.clear();
-const shelf = []
-const localStorageKey = 'bookFinder_bookShelf';
-if (!localStorage.getItem(localStorageKey)) {
-    localStorage.setItem(localStorageKey, JSON.stringify(shelf));
-  }
-const localBookshelf = JSON.parse(localStorage.getItem(localStorageKey));
-
-const bookCards = localBookshelf.map(item => 
+class Bookshelf extends Component {
+  render() {
+    const {savedBooks} = this.props;
+    const bookCards = savedBooks.map(item => 
                   <Cards key={item.id} 
                         id={item.id} 
                         bookTitle={item.title} 
@@ -20,16 +14,8 @@ const bookCards = localBookshelf.map(item =>
                         bookLink={item.bookLink}
                         imageLink={item.image}
                         saved={true}
+                        savedBooks={savedBooks}
                   />);
-class Bookshelf extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      
-    };
-  }
-
-  render() {
     return (
       <div>
         <Route render={({history}) => (
